more tests, better dialogs
