+++ frontends/infinite-corridor/src/pages/user-lists/UserListsListingPage.tsx
@@ -17,7 +17,8 @@ import {
   useDeletionDialog,
   useEditingDialog,
   useCreationDialog
-} from "./ManageListDialog"
+} from "./ManageListDialogs"
+import UpsertListDialog from "./UpsertListDialog"
 import { GridColumn, GridContainer } from "../../components/layout"
 import { useFavorites, useUserListsListing } from "../../api/learning-resources"
 import Container from "@mui/material/Container"
