+++ frontends/ol-util/src/test-utils/index.ts
@@ -87,6 +87,19 @@ const getDescriptionFor = (el: HTMLElement) => {
   return errEl
 }
 
+/**
+ * We use [jest-fail-on-console](https://www.npmjs.com/package/jest-fail-on-console)
+ * to fail on console errors/warnings.
+ *
+ * This function allows us to temporarily disable that behavior for a test,
+ * e.g., to test API error handling.
+ */
+const allowConsoleErrors = () => {
+  const consoleError = jest.spyOn(console, "error").mockImplementation()
+  const consoleWarn = jest.spyOn(console, "warn").mockImplementation()
+  return { consoleError, consoleWarn }
+}
+
 export {
   queryAllByTerm,
   queryByTerm,
@@ -95,5 +108,6 @@ export {
   findAllByTerm,
   findByTerm,
   getDescriptionFor,
-  createMatchMediaForJsDom
+  createMatchMediaForJsDom,
+  allowConsoleErrors
 }
